+++ lib/database/lib/model/db-node-summary.ts
@@ -62,7 +62,7 @@ export interface DbNodeSummary {
    */
   "softwareStorageSizeInGB"?: number;
   /**
-   * The type of maintenance of dbNode.
+   * The type of database node maintenance.
    */
   "maintenanceType"?: DbNodeSummary.MaintenanceType;
   /**
@@ -74,7 +74,7 @@ export interface DbNodeSummary {
    */
   "timeMaintenanceWindowEnd"?: Date;
   /**
-   * Additional information like a message to customer about the maintenance.
+   * Additional information about the planned maintenance.
    */
   "additionalDetails"?: string;
 }