+++ lib/identity/lib/model/create-network-source-details.ts
@@ -19,7 +19,7 @@ import common = require("oci-common");
  */
 export interface CreateNetworkSourceDetails {
   /**
-   * The OCID of the tenancy containing the network source object.
+   * The OCID of the tenancy (root compartment) containing the network source object.
    */
   "compartmentId": string;
   /**
@@ -29,18 +29,20 @@ export interface CreateNetworkSourceDetails {
    */
   "name": string;
   /**
-   * A list of allowed public IPs and CIDR ranges
+   * A list of allowed public IP addresses and CIDR ranges.
    *
    */
   "publicSourceList"?: Array<string>;
   /**
-   * A list of allowed VCN ocid/IP range pairs
+   * A list of allowed VCN OCID and IP range pairs.
+   * Example:`\"vcnId\": \"ocid1.vcn.oc1.iad.aaaaaaaaexampleuniqueID\", \"ipRanges\": [ \"129.213.39.0/24\" ]`
    *
    */
   "virtualSourceList"?: Array<model.NetworkSourcesVirtualSourceList>;
   /**
-   * A list of OCIservices allowed to make on behalf of requests which may have different source ips.
-   * At this time only the values of all or none are supported.
+   * A list of services allowed to make on-behalf-of requests. These requests can have different source IP addresses
+   * than those listed in the network source.
+   * Currently, only `all` and `none` are supported. The default is `all`.
    *
    */
   "services"?: Array<string>;