+++ lib/identity/lib/model/create-region-subscription-details.ts
@@ -16,16 +16,8 @@ import common = require("oci-common");
 
 export interface CreateRegionSubscriptionDetails {
   /**
-    * The regions's key.
-* <p>
-Allowed values are:
-* - `PHX`
-* - `IAD`
-* - `FRA`
-* - `LHR`
-* - `YYZ`
-* - `NRT`
-* - `ICN`
+    * The regions's key. See [Regions and Availability Domains](https://docs.cloud.oracle.com/Content/General/Concepts/regions.htm) for
+* the full list of supported 3-letter region codes.
 * <p>
 Example: `PHX`
 *