+++ lib/identity/lib/model/network-sources-summary.ts
@@ -15,8 +15,9 @@ import * as model from "../model";
 import common = require("oci-common");
 
 /**
- * A network source defines a list of source IPs that are allowed to make auth requests
- * More info needed here
+ * A network source specifies a list of source IP addresses that are allowed to make authorization requests.
+ * Use the network source in policy statements to restrict access to only requests that come from the specified IPs.
+ * For more information, see [Managing Network Sources](https://docs.cloud.oracle.com/Content/Identity/Tasks/managingnetworksources.htm).
  *
  */
 export interface NetworkSourcesSummary {
@@ -25,7 +26,7 @@ export interface NetworkSourcesSummary {
    */
   "id"?: string;
   /**
-   * The OCID of the tenancy containing the network source.
+   * The OCID of the tenancy (root compartment) containing the network source.
    */
   "compartmentId"?: string;
   /**
@@ -39,18 +40,19 @@ export interface NetworkSourcesSummary {
    */
   "description"?: string;
   /**
-   * A list of allowed public IPs and CIDR ranges
+   * A list of allowed public IP addresses and CIDR ranges.
    *
    */
   "publicSourceList"?: Array<string>;
   /**
-   * A list of allowed VCN ocid/IP range pairs
+   * A list of allowed VCN OCID and IP range pairs.
+   * Example:`\"vcnId\": \"ocid1.vcn.oc1.iad.aaaaaaaaexampleuniqueID\", \"ipRanges\": [ \"129.213.39.0/24\" ]`
    *
    */
   "virtualSourceList"?: Array<model.NetworkSourcesVirtualSourceList>;
   /**
-   * A list of OCIservices allowed to make on behalf of requests which may have different source ips.
-   * At this time only the values of all or none are supported.
+   * A list of services allowed to make on-behalf-of requests. These requests can have different source IPs than
+   * those specified in the network source. Currently, only `all` and `none` are supported. The default is `all`.
    *
    */
   "services"?: Array<string>;