+++ lib/integration/lib/model/create-integration-instance-details.ts
@@ -56,6 +56,10 @@ export interface CreateIntegrationInstanceDetails {
    * The number of configured message packs
    */
   "messagePacks": number;
+  /**
+   * The file server is enabled or not.
+   */
+  "isFileServerEnabled"?: boolean;
 }
 
 export namespace CreateIntegrationInstanceDetails {