+++ lib/integration/lib/model/integration-instance-summary.ts
@@ -62,6 +62,10 @@ export interface IntegrationInstanceSummary {
    * The number of configured message packs (if any)
    */
   "messagePacks": number;
+  /**
+   * The file server is enabled or not.
+   */
+  "isFileServerEnabled"?: boolean;
 }
 
 export namespace IntegrationInstanceSummary {