+++ CHANGELOG.md
@@ -4,7 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## 1.0.2 - 20-06-02
+## 1.1.0 - 2020-06-09
+
+### Added
+
+- Support for returning the database version of backups in the Database service
+- Support for patching on Exadata Cloud at Customer resources in the Database service
+- Support for new lifecycle substates on instances in the Digital Assistant service
+- Support for file servers in the Integration service
+- Support for deleting non-empty tag namespaces and bulk deleting tags in the Identity service
+- Support for bulk move and bulk delete of resources by compartment in the Identity service
+
+### Breaking
+
+- Parameter LifecycleState removed state OFFLINE and added DISCONNECTED in the Database service
+
+## 1.0.2 - 2020-06-02
 
 ### Added